Added common functions for locating and managing sensor data.
Functions include functions to aid in the location of sensor data files
on the ROGER file system, extraction of site boundary data and conversion
to geojson, and clipping of sensor data to boundary files.

This also introduced dependency on GDAL and PIL modules. GDAL seems
problematic in that the module must be installed at the system level
(ie no virtualenv) and even then it doesn't work for python3.