Improve handling of BETYDB_URL and BETYDB_KEY.
Started the transition to using environmental variable to pass in the
location of betydb and the necessary key. This will needs to be made
more robust.

Most of the utility functions (ie betydb_traits, etc) didn't work
correctly that has been fixed but it still need more work. Some functions
don't work because they don't allow the passing of betykey parameter.