Remove the use of the level parameter and changed create_sensor_path.
Modify the create_sensor_path to match get_sensor_path in
terms of parameters but have a side effect of creating necessary
path directories.

Fix issue with submit_traits in betydb.py where it was using local
functions to set default parameter values. This had a side effect
of calling the functions during import which broke testing and
is bad practice in general.